prytkin_timofey_lab_2 is ready
