+++ WindowsFormsAtackAircraft/WindowsFormsAtackAircraft/ITransport.cs
@@ -1,0 +1,30 @@
+﻿using System.Drawing;
+
+
+namespace WindowsFormsAtackAircraft
+{
+    interface ITransport
+
+    {
+        /// <summary>
+        /// Установка позиции
+        /// </summary>
+        /// <param name="x">Координата X</param>
+        /// <param name="y">Координата Y</param>
+        /// <param name="width">Ширина картинки</param>
+        /// <param name="height">Высота картинки</param>
+        void SetPosition(int x, int y, int width, int height);
+
+        /// <summary>
+        /// Изменение направления пермещения
+        /// </summary>
+        /// <param name="direction">Направление</param>
+        void MoveTransport(Direction direction);
+
+        /// <summary>
+        /// Отрисовка
+        /// </summary>
+        /// <param name="g"></param>
+        void DrawTransport(Graphics g);
+    }
+}
