+++ WindowsFormsAtackAircraft/WindowsFormsAtackAircraft/WindowsFormsAtackAircraft.csproj
@@ -46,14 +46,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AtackAircraft.cs" />
+    <Compile Include="AttackAircraft.cs" />
     <Compile Include="Direction.cs" />
+    <Compile Include="FlyingTransport.cs" />
     <Compile Include="FormAtackAircraft.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="FormAtackAircraft.Designer.cs">
       <DependentUpon>FormAtackAircraft.cs</DependentUpon>
     </Compile>
+    <Compile Include="ITransport.cs" />
+    <Compile Include="Plane.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="FormAtackAircraft.resx">
